+++ packages/DataHeader/stories/DataHeader.withTypes.stories.js
@@ -6,7 +6,7 @@ import EllipsisVertical from "@paprika/icon/lib/EllipsisVertical";
 
 import { Story } from "storybook/assets/styles/common.styles";
 import { getStoryName } from "storybook/storyTree";
-import DropdownMenu from "@paprika/dropdown-menu";
+import OverflowMenu from "@paprika/overflow-menu";
 import DataHeader from "../src";
 
 export default {
@@ -26,14 +26,14 @@ function Menu(props) {
   };
 
   return (
-    <DropdownMenu>
-      <DropdownMenu.Trigger buttonType="raw">
+    <OverflowMenu>
+      <OverflowMenu.Trigger buttonType="raw">
         <EllipsisVertical />
-      </DropdownMenu.Trigger>
-      <DropdownMenu.Item onClick={handleClick("one")}>One</DropdownMenu.Item>
-      <DropdownMenu.Item onClick={handleClick("two")}>Two</DropdownMenu.Item>
-      <DropdownMenu.Item onClick={handleClick("three")}>Three</DropdownMenu.Item>
-    </DropdownMenu>
+      </OverflowMenu.Trigger>
+      <OverflowMenu.Item onClick={handleClick("one")}>One</OverflowMenu.Item>
+      <OverflowMenu.Item onClick={handleClick("two")}>Two</OverflowMenu.Item>
+      <OverflowMenu.Item onClick={handleClick("three")}>Three</OverflowMenu.Item>
+    </OverflowMenu>
   );
 }