+++ packages/CollapsibleCard/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Replace `nanoid` by `uuid`.
+
 ## [1.1.2-alpha.1] - 2020-12-04
 
 - Support render function for `label` prop.